requirements.txt
# Core ML
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Web App
streamlit>=1.51.0
streamlit-webrtc>=0.47.0

# Data & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0
opencv-python-headless>=4.8.0

# Video Processing
moviepy>=1.0.3
av>=10.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0